Prevent unsafe narrowing: base/ for Mac
Bug: 1292951
No-Presubmit: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3752846
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1023512}

+++ base/process/launch_mac.cc
@@ -153,7 +153,7 @@ bool GetAppOutputInternal(const std::vector<std::string>& argv,
     read_this_pass = HANDLE_EINTR(
         read(read_fd.get(), &(*output)[total_bytes_read], kBufferSize));
     if (read_this_pass >= 0) {
-      total_bytes_read += read_this_pass;
+      total_bytes_read += static_cast<size_t>(read_this_pass);
       output->resize(total_bytes_read);
     }
   } while (read_this_pass > 0);
